requirements.txt
numpy>=1.20.0
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
redis>=4.0.0
